serializer.rs, go.rs: "Fix" `},{` collapsing code and update Go tests to match
"Fix" because it is still very broken:

  a) it's using hardcoded characters instead of the settings in Langauge
  b) it doesn't really have the correct information at this point in the code

Not going to fix a since b is unfixable. Instead, we should move this to the
flush() routine where it can live as sort of a peephole optimizer. Ideally it
should be able to figure out:
  [
    { ..small }, { ..small },
    {
      ..large
    }, {
      ..largs
    },
    { ..med },
    { ..med }
  ]
Currently it can't do that because it can't decide of things are small or
large--flush() can.